+++ bin/proxyl.rb
@@ -3,12 +3,12 @@
 require 'tty-spinner'
 require 'colorize'
 require 'http'
-require_relative '../lib/scrapper'
+require_relative '../lib/scraper'
 require_relative '../lib/file_manager'
 
 spinner = TTY::Spinner.new(':spinner :title', format: :bouncing, hide_cursor: true)
 document = nil
-scrapper = nil
+scraper = nil
 file = nil
 
 spinner.update(title: 'Downloading data...')
@@ -18,16 +18,16 @@ spinner.run('Done!') do |s|
   s.reset
 end
 
-spinner.update(title: 'Scrapping...')
+spinner.update(title: 'Scraping...')
 spinner.run('Done!') do |s|
-  scrapper = Scrapper.new(document)
-  scrapper.scrape
+  scraper = Scraper.new(document)
+  scraper.scrape
   s.reset
 end
 
 spinner.update(title: 'Saving...')
 spinner.run('Done!') do |s|
-  file = FileManager.save(scrapper.data.to_json, 'json')
+  file = FileManager.save(scraper.data.to_json, 'json')
   s.reset
 end